chore: convert project to monorepo
