+++ fixtures/fancy-accordion/tsconfig.json
@@ -8,7 +8,11 @@
     "rootDir": "./src",
     "composite": true
   },
-  "references": [],
+  "references": [
+    {
+      "path": "../expansion-panel/tsconfig.json"
+    }
+  ],
   "include": [
     "src"
   ],
