+++ package.json
@@ -66,16 +66,6 @@
     "update:tsconfigs": "node scripts/generate-ts-configs.js",
     "watch": "tsc-watch"
   },
-  "dependencies": {
-    "@types/dompurify": "^2.3.1",
-    "@types/marked": "^4.0.0",
-    "custom-elements-manifest": "^1.0.0",
-    "dompurify": "^2.3.3",
-    "highlight-ts": "9.12.1-2",
-    "lit": "^2.0.0",
-    "marked": "^4.0.0",
-    "tslib": "^2.3.1"
-  },
   "devDependencies": {
     "@custom-elements-manifest/analyzer": "^0.5.7",
     "@rocket/cli": "^0.10.1",
