+++ packages/api-common/src/manifest.ts
