+++ packages/api-common/src/templates.ts
