+++ packages/api-demo/src/styles.ts
@@ -1,5 +1,63 @@
 import { css } from 'lit';
-import highlightTheme from './highlight-styles.js';
+
+const highlightTheme = css`
+  pre {
+    margin: 0;
+    color: black;
+    background: none;
+    font-family: var(--ave-monospace-font);
+    font-size: 0.875rem;
+    text-align: left;
+    white-space: pre-wrap;
+    word-spacing: normal;
+    word-break: normal;
+    word-wrap: normal;
+    line-height: 1.5;
+    tab-size: 4;
+    hyphens: none;
+    text-shadow: none;
+  }
+
+  code {
+    font-family: inherit;
+  }
+
+  .comment {
+    color: slategray;
+  }
+
+  .attr,
+  .selector-tag {
+    color: #690;
+  }
+
+  .css {
+    color: #333;
+  }
+
+  .built_in {
+    color: #dd4a68;
+  }
+
+  .meta {
+    color: #e90;
+    font-weight: bold;
+  }
+
+  .string {
+    color: #07a;
+  }
+
+  .tag {
+    color: #999;
+  }
+
+  .attribute,
+  .name,
+  .number {
+    color: #905;
+  }
+`;
 
 const demoStyles = css`
   button {
@@ -19,7 +77,7 @@ const demoStyles = css`
     background: var(--ave-button-active-background, rgba(0, 0, 0, 0.6));
   }
 
-  api-viewer-demo {
+  api-demo-layout {
     display: block;
   }
