+++ packages/api-demo/tsconfig.json
@@ -8,7 +8,11 @@
     "rootDir": "./src",
     "composite": true
   },
-  "references": [],
+  "references": [
+    {
+      "path": "../api-tabs/tsconfig.json"
+    }
+  ],
   "include": [
     "src"
   ],
