+++ packages/api-docs/package.json
@@ -44,11 +44,11 @@
     "web-components"
   ],
   "dependencies": {
+    "@api-viewer/common": "^0.1.0",
+    "@api-viewer/tabs": "^0.1.0",
     "@types/dompurify": "^2.3.1",
     "@types/marked": "^4.0.0",
-    "custom-elements-manifest": "^1.0.0",
     "dompurify": "^2.3.3",
-    "highlight-ts": "9.12.1-2",
     "lit": "^2.0.0",
     "marked": "^4.0.0",
     "tslib": "^2.3.1"
