+++ packages/api-docs/src/api-docs.ts
@@ -1,6 +1,6 @@
-import { ApiDocsBase } from './api-docs-base.js';
-import docsStyles from './styles/docs-styles.js';
-import sharedStyles from './styles/shared-styles.js';
+import sharedStyles from '@api-viewer/common/lib/shared-styles.js';
+import { ApiDocsBase } from './base.js';
+import docsStyles from './styles.js';
 
 export class ApiDocs extends ApiDocsBase {
   static get styles() {
