+++ packages/api-tabs/tsconfig.json
@@ -8,17 +8,7 @@
     "rootDir": "./src",
     "composite": true
   },
-  "references": [
-    {
-      "path": "../api-tabs/tsconfig.json"
-    },
-    {
-      "path": "../api-demo/tsconfig.json"
-    },
-    {
-      "path": "../api-docs/tsconfig.json"
-    }
-  ],
+  "references": [],
   "include": [
     "src"
   ],
