--- src/styles/highlight-styles.ts
@@ -1,60 +1,0 @@
-import { css } from 'lit';
-
-export default css`
-  pre {
-    margin: 0;
-    color: black;
-    background: none;
-    font-family: var(--ave-monospace-font);
-    font-size: 0.875rem;
-    text-align: left;
-    white-space: pre-wrap;
-    word-spacing: normal;
-    word-break: normal;
-    word-wrap: normal;
-    line-height: 1.5;
-    tab-size: 4;
-    hyphens: none;
-    text-shadow: none;
-  }
-
-  code {
-    font-family: inherit;
-  }
-
-  .comment {
-    color: slategray;
-  }
-
-  .attr,
-  .selector-tag {
-    color: #690;
-  }
-
-  .css {
-    color: #333;
-  }
-
-  .built_in {
-    color: #dd4a68;
-  }
-
-  .meta {
-    color: #e90;
-    font-weight: bold;
-  }
-
-  .string {
-    color: #07a;
-  }
-
-  .tag {
-    color: #999;
-  }
-
-  .attribute,
-  .name,
-  .number {
-    color: #905;
-  }
-`;
