+++ workspace-packages.js
@@ -1,6 +1,8 @@
 const packages = [
+  { name: 'api-common', type: 'ts', environment: 'browser' },
   { name: 'api-demo', type: 'ts', environment: 'browser' },
   { name: 'api-docs', type: 'ts', environment: 'browser' },
+  { name: 'api-tabs', type: 'ts', environment: 'browser' },
   { name: 'api-viewer', type: 'ts', environment: 'browser' },
   { name: 'expansion-panel', type: 'ts', environment: 'browser' },
   { name: 'fancy-accordion', type: 'ts', environment: 'browser' },
